Rewritten expression parser.
No more RPN!
The new version is waaaay better and about 250 times faster (well, at
least on my computer).
No code generator changes though.

+++ optimization/code/opt_constant_folding.pas
@@ -6,7 +6,7 @@ Procedure __constant_folding(const ErrorOnInvalidOperator: Boolean);
 Procedure Parse(var Expr: PExpressionNode);
 Var Tmp, Tmp2, Left, Right: PExpressionNode;
     Evaluated             : Boolean = False;
-    I                     : Integer;
+    I                     : int16;
 Begin
  if (Expr = nil) Then // nothing to do
   Exit;
